# apps/be/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/catalog-nodes/{id}": {
      "get": {
        "operationId": "CatalogNodes_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNode"
                }
              }
            }
          }
        },
        "summary": "Get catalog node by id",
        "tags": [
          "Catalog - nodes"
        ]
      },
      "patch": {
        "operationId": "CatalogNodes_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNode"
                }
              }
            }
          }
        },
        "summary": "Update catalog node",
        "tags": [
          "Catalog - nodes"
        ]
      },
      "delete": {
        "description": "Hard delete; cascades to versions when no FKs block it.",
        "operationId": "CatalogNodes_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node type",
        "tags": [
          "Catalog - nodes"
        ]
      }
    },
...
    "/catalog-node-versions/{catalogNodeVersionId}/sockets": {
      "get": {
        "operationId": "CatalogNodeSocketsList_findAll",
        "parameters": [
          {
            "name": "catalogNodeVersionId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "input",
                "output"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCatalogNodeSockets"
                }
              }
            }
          }
        },
        "summary": "List sockets on a catalog version (paginated)",
        "tags": [
          "Catalog - node sockets"
        ]
      }
    },
...
    "/catalog-node-versions/{catalogNodeVersionId}/socket-rules": {
      "get": {
        "operationId": "CatalogNodeSocketRulesList_findAll",
        "parameters": [
          {
            "name": "catalogNodeVersionId",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Column to sort by (resource-specific whitelist)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "default": "ASC",
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Text search (maps to searchable columns per endpoint)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedCatalogNodeSocketRules"
                }
              }
            }
          }
        },
        "summary": "List socket rules on a catalog version (paginated)",
        "tags": [
          "Catalog - socket rules"
        ]
      }
    },
...
    "/catalog-node-sockets/{id}": {
      "get": {
        "operationId": "CatalogNodeSockets_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocket"
                }
              }
            }
          }
        },
        "summary": "Get catalog node socket by id",
        "tags": [
          "Catalog - node sockets"
        ]
      },
      "patch": {
        "operationId": "CatalogNodeSockets_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeSocketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocket"
                }
              }
            }
          }
        },
        "summary": "Update catalog node socket",
        "tags": [
          "Catalog - node sockets"
        ]
      },
      "delete": {
        "operationId": "CatalogNodeSockets_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node socket",
        "tags": [
          "Catalog - node sockets"
        ]
      }
    },
    "/catalog-node-socket-rules": {
      "post": {
        "operationId": "CatalogNodeSocketRules_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogNodeSocketRuleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocketRule"
                }
              }
            }
          }
        },
        "summary": "Create allowed socket pair on a catalog version",
        "tags": [
          "Catalog - socket rules"
        ]
      }
    },
    "/catalog-node-socket-rules/{id}": {
      "get": {
        "operationId": "CatalogNodeSocketRules_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocketRule"
                }
              }
            }
          }
        },
        "summary": "Get catalog node socket rule by id",
        "tags": [
          "Catalog - socket rules"
        ]
      },
      "patch": {
        "operationId": "CatalogNodeSocketRules_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodeSocketRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeSocketRule"
                }
              }
            }
          }
        },
        "summary": "Update catalog node socket rule",
        "tags": [
          "Catalog - socket rules"
        ]
      },
      "delete": {
        "operationId": "CatalogNodeSocketRules_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node socket rule",
        "tags": [
          "Catalog - socket rules"
        ]
      }
    },
    "/catalog-node-properties": {
      "post": {
        "operationId": "CatalogNodeProperties_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCatalogNodePropertyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeProperty"
                }
              }
            }
          }
        },
        "summary": "Create property schema on a catalog version",
        "tags": [
          "Catalog - node properties"
        ]
      }
    },
    "/catalog-node-properties/{id}": {
      "get": {
        "operationId": "CatalogNodeProperties_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeProperty"
                }
              }
            }
          }
        },
        "summary": "Get catalog node property by id",
        "tags": [
          "Catalog - node properties"
        ]
      },
      "patch": {
        "operationId": "CatalogNodeProperties_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCatalogNodePropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogNodeProperty"
                }
              }
            }
          }
        },
        "summary": "Update catalog node property",
        "tags": [
          "Catalog - node properties"
        ]
      },
      "delete": {
        "operationId": "CatalogNodeProperties_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Row primary key (`bigint`, string in JSON)",
            "schema": {
              "type": "string",
              "example": "1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": ""
          }
        },
        "summary": "Delete catalog node property",
        "tags": [
          "Catalog - node properties"
        ]
      }
    },
...
  },
  "info": {
    "title": "Shake API",
    "description": "REST API for the **node catalog** (toolbox definitions) and **board graph** (runtime nodes, connections, props).\n\nEntity IDs are `bigint` in PostgreSQL and serialized as **strings** in JSON.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CatalogNode": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "slug": {
            "type": "string",
            "example": "http-request",
            "maxLength": 128
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "slug"
        ]
      },
      "PaginatedMetaDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "minimum": 1
          },
          "pageSize": {
            "type": "number",
            "example": 20,
            "minimum": 1,
            "maximum": 100
          },
          "total": {
            "type": "number",
            "example": 142,
            "minimum": 0
          },
          "totalPages": {
            "type": "number",
            "example": 8,
            "minimum": 0
          }
        },
        "required": [
          "page",
          "pageSize",
          "total",
          "totalPages"
        ]
      },
...
      "UpdateCatalogNodeDto": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string",
            "maxLength": 128,
            "example": "tile"
          }
        }
      },
...
      "CatalogNodeSocket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "example": "input"
          },
          "name": {
            "type": "string",
            "example": "body",
            "maxLength": 255
          },
          "limit": {
            "type": "object",
            "example": 1,
            "nullable": true
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeVersionId",
          "type",
          "name"
        ]
      },
      "PaginatedCatalogNodeSockets": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogNodeSocket"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CatalogNodeProperty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "type": {
            "type": "string",
            "example": "string",
            "maxLength": 64
          },
          "defaultValue": {
            "type": "object",
            "nullable": true
          },
          "isRequired": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeVersionId",
          "type",
          "isRequired"
        ]
      },
...
      "CatalogNodeSocketRule": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "updatedById": {
            "type": "string",
            "example": "1",
            "nullable": true,
            "description": "bigint (string in JSON)"
          },
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodeSocketFromId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          },
          "catalogNodeSocketToId": {
            "type": "string",
            "example": "1",
            "description": "bigint (string in JSON)"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "catalogNodeVersionId",
          "catalogNodeSocketFromId",
          "catalogNodeSocketToId"
        ]
      },
      "PaginatedCatalogNodeSocketRules": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogNodeSocketRule"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PaginatedMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
      "UpdateCatalogNodeSocketDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "type": {
            "type": "string",
            "enum": [
              "input",
              "output"
            ],
            "example": "input"
          },
          "name": {
            "type": "string",
            "maxLength": 255,
            "example": "children"
          },
          "limit": {
            "type": "number",
            "nullable": true,
            "description": "Max connections allowed to this socket (null = unlimited)"
          }
        }
      },
      "CreateCatalogNodeSocketRuleDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "catalogNodeSocketFromId": {
            "type": "string",
            "example": "10",
            "description": "FK to catalog_node_sockets.id"
          },
          "catalogNodeSocketToId": {
            "type": "string",
            "example": "11",
            "description": "FK to catalog_node_sockets.id"
          }
        },
        "required": [
          "catalogNodeVersionId",
          "catalogNodeSocketFromId",
          "catalogNodeSocketToId"
        ]
      },
      "UpdateCatalogNodeSocketRuleDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "catalogNodeSocketFromId": {
            "type": "string",
            "example": "10",
            "description": "FK to catalog_node_sockets.id"
          },
          "catalogNodeSocketToId": {
            "type": "string",
            "example": "11",
            "description": "FK to catalog_node_sockets.id"
          }
        }
      },
      "CreateCatalogNodePropertyDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "type": {
            "type": "string",
            "maxLength": 64,
            "example": "text",
            "description": "e.g. `text`, `integer`"
          },
          "defaultValue": {
            "type": "object",
            "description": "JSON value; shape depends on `type`",
            "example": "hello"
          },
          "isRequired": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "catalogNodeVersionId",
          "type"
        ]
      },
      "UpdateCatalogNodePropertyDto": {
        "type": "object",
        "properties": {
          "catalogNodeVersionId": {
            "type": "string",
            "example": "1",
            "description": "FK to catalog_node_versions.id"
          },
          "type": {
            "type": "string",
            "maxLength": 64,
            "example": "text",
            "description": "e.g. `text`, `integer`"
          },
          "defaultValue": {
            "type": "object",
            "description": "JSON value; shape depends on `type`",
            "example": "hello"
          },
          "isRequired": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
    }
  }
}